+++ crates/iota-core/src/authority/consensus_quarantine.rs
@@ -78,6 +78,10 @@ impl ConsensusCommitOutput {
         self.next_randomness_round.as_ref().map(|(_, ts)| *ts)
     }
 
+    pub fn insert_end_of_publish(&mut self, authority: AuthorityName, partial_scores: Vec<u32>) {
+        self.end_of_publish.insert(authority);
+    }
+
     fn get_highest_pending_checkpoint_height(&self) -> Option<CheckpointHeight> {
         self.pending_checkpoints.last().map(|cp| cp.height())
     }
