+++ crates/iota-core/src/consensus_manager/mysticeti_manager.rs
@@ -185,7 +185,6 @@ impl ConsensusManagerTrait for MysticetiManager {
                 *boot_counter
             );
         }
-        let partial_scores = epoch_store.partial_scores.clone();
         let authority = ConsensusAuthority::start(
             network_type,
             own_index,
@@ -197,7 +196,7 @@ impl ConsensusManagerTrait for MysticetiManager {
             Arc::new(tx_validator.clone()),
             consumer,
             registry.clone(),
-            partial_scores,
+            epoch_store.scorer.clone(),
             *boot_counter,
         )
         .await;
